Removed some compiler warnings and bumped the About date.
git-svn-id: svn://scribus.net/branches/Version135x/Scribus@9657 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/fonts/scface_ttf.cpp
@@ -44,9 +44,9 @@ bool copy(QByteArray & dst, uint to, QByteArray & src, uint from, uint len)
 		return false;
 	if (!src.data())
 		return false;
-	if (to + len > dst.size())
+	if (to + len > static_cast<uint>(dst.size()))
 		return false;
-	if (from + len > src.size())
+	if (from + len > static_cast<uint>(src.size()))
 		return false;
 	
 	memcpy(dst.data() + to, src.data() + from, len);
